+++ config/me_instagram.php
@@ -9,7 +9,7 @@ return [
         //Frontend
         'frontend' => [
             //Number of photos to show per page. This must be a multiple of 3
-            'photos' => 15,
+            'photos' => 12,
             //Show the "follow me button in the user's profile
             'follow_me' => TRUE,
             //Open the photos on Instagram, rather than on the site