+++ src/Utility/Sitemap.php
@@ -1,0 +1,46 @@
+<?php
+/**
+ * This file is part of MeInstagram.
+ *
+ * MeInstagram is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * MeInstagram is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with MeInstagram.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ * @author		Mirko Pagliai <mirko.pagliai@gmail.com>
+ * @copyright	Copyright (c) 2016, Mirko Pagliai for Nova Atlantis Ltd
+ * @license		http://www.gnu.org/licenses/agpl.txt AGPL License
+ * @link		http://git.novatlantis.it Nova Atlantis Ltd
+ * @see         MeCms\Utility\SitemapBuilder
+ */
+namespace MeInstagram\Utility;
+
+use MeCms\Utility\SitemapBuilder;
+
+/**
+ * This class contains methods called by the `SitemapBuilder`.
+ * Each method must be return an array or urls to add to the sitemap.
+ */
+class Sitemap extends SitemapBuilder {
+    /**
+     * Method that returns instagram urls
+     * @return array
+     * @uses MeCms\Utility\SitemapBuilder::url()
+     */
+    public static function instagram() {
+        //Adds Instagram index
+        $url = [
+            self::url(['_name' => 'instagram']),
+        ];
+        
+        return $url;
+    }
+}